+++ aea/package_manager/__init__.py
@@ -1,8 +1,7 @@
 # -*- coding: utf-8 -*-
 # ------------------------------------------------------------------------------
 #
-#   Copyright 2021-2022 Valory AG
-#   Copyright 2018-2020 Fetch.AI Limited
+#   Copyright 2022 Valory AG
 #
 #   Licensed under the Apache License, Version 2.0 (the "License");
 #   you may not use this file except in compliance with the License.
@@ -18,21 +17,4 @@
 #
 # ------------------------------------------------------------------------------
 
-"""Module with exceptions of the aea cli."""
-
-from warnings import warn
-
-import click
-
-from aea.exceptions import AEAException
-
-
-def aev_flag_depreaction() -> None:
-    """Deprecation warning for `--aev` flag."""
-    message = "`--aev` flag is deprecated and will be removed in v2.0.0, usage of envrionment varibales is default now."
-    click.echo(message)
-    warn(message, DeprecationWarning, stacklevel=2)
-
-
-class AEAConfigException(AEAException):
-    """Exception about AEA configuration."""
+"""AEA Package manager."""